ModemConfig: Register subchange with SubscriptionManager instead
DEFAULT_SUBSCRIPTION_CHANGED is not delivered often enough. When
swapping the same SIM between slots the selected subscription doesn't
change yet the firmware for the slot might need to be updated.
Registering this handler provides the ability to do so, but has to be
registered from a background service that consequently is to be started
by a BootReceiver.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>