ModemConfig: Convert getProviders one-time-init to lazyinit
Lazyinit allows a non-nullable to be null "temporarily" while it is
initialized outside of a constructor. This is useful in android
scenarios where initialization usually happens in onCreate().

Note that it is not really safe. The compiler doesn't do any checks or
use annotations to understand that this value should be initialized in
onCreate, nor does it fail compilation if it is never initialized at
all or used before initialization. After all `providers` remains a
`var`iable, not a `val`ue. In hindsight `lateinit` is simply a free pass
to circumvent compile-time null checks, and is best avoided.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>